☂️ Complete the Tailwind v4 setup
