+++ lib/logr_web/components/core_components.ex
@@ -581,7 +581,7 @@ defmodule LogrWeb.CoreComponents do
   width, height, and background color classes.
 
   Icons are extracted from the `deps/heroicons` directory and bundled within
-  your compiled app.css by the plugin in your `assets/tailwind.config.js`.
+  your compiled app.css by the plugin in your `assets/css/tailwind_heroicons.js`.
 
   ## Examples
